don't need to care about long decimals

+++ src/main/java/lol/hyper/ezhomes/commands/CommandWhere.java
@@ -33,9 +33,9 @@ public class CommandWhere implements CommandExecutor {
                             sender.sendMessage(ChatColor.GOLD + "--------------------------------------------");
                             sender.sendMessage(ChatColor.GOLD + args[0] + "'s location:");
                             sender.sendMessage(ChatColor.YELLOW + "World: " + ChatColor.GOLD + home.getWorld().getName());
-                            sender.sendMessage(ChatColor.YELLOW + "X: " + ChatColor.GOLD + home.getX());
-                            sender.sendMessage(ChatColor.YELLOW + "Y: " + ChatColor.GOLD + home.getY());
-                            sender.sendMessage(ChatColor.YELLOW + "Y: " + ChatColor.GOLD + home.getZ());
+                            sender.sendMessage(ChatColor.YELLOW + "X: " + ChatColor.GOLD + (int) home.getX());
+                            sender.sendMessage(ChatColor.YELLOW + "Y: " + ChatColor.GOLD + (int) home.getY());
+                            sender.sendMessage(ChatColor.YELLOW + "Y: " + ChatColor.GOLD + (int) home.getZ());
                             sender.sendMessage(ChatColor.GOLD + "--------------------------------------------");
                         } else {
                             sender.sendMessage(ChatColor.RED + "That home does not exist.");
